refactor(observer): move bindings to the related widgets
