+++ src/widgets/label/lv_label.h
@@ -23,6 +23,7 @@ extern "C" {
 #include "../../font/lv_symbol_def.h"
 #include "../../misc/lv_text.h"
 #include "../../draw/lv_draw.h"
+#include "../../others/observer/lv_observer.h"
 
 /*********************
  *      DEFINES
@@ -227,6 +228,22 @@ bool lv_label_get_recolor(const lv_obj_t * obj);
  * Other functions
  *====================*/
 
+#if LV_USE_OBSERVER
+/**
+ * Bind an integer, string, or pointer Subject to a Label.
+ * @param obj       pointer to Label
+ * @param subject   pointer to Subject
+ * @param fmt       optional printf-like format string with 1 format specifier (e.g. "%d °C")
+ *                  or NULL to bind to the value directly.
+ * @return          pointer to newly-created Observer
+ * @note            `fmt == NULL` can be used only with string and pointer Subjects.
+ * @note            If Subject is a pointer and `fmt == NULL`, pointer must point
+ *                  to a `\0` terminated string.
+ */
+lv_observer_t * lv_label_bind_text(lv_obj_t * obj, lv_subject_t * subject, const char * fmt);
+#endif
+
+
 /**
  * Insert a text to a label. The label text cannot be static.
  * @param obj       pointer to a label object
@@ -244,6 +261,8 @@ void lv_label_ins_text(lv_obj_t * obj, uint32_t pos, const char * txt);
  * @param cnt       number of characters to cut
  */
 void lv_label_cut_text(lv_obj_t * obj, uint32_t pos, uint32_t cnt);
+
+
 
 /**********************
  *      MACROS
