+++ src/widgets/slider/lv_slider.c
@@ -22,6 +22,7 @@
 #include "../../stdlib/lv_string.h"
 #include "../../misc/lv_math.h"
 #include "../image/lv_image.h"
+#include "../../others/observer/lv_observer_private.h"
 
 /*********************
  *      DEFINES
@@ -45,6 +46,11 @@ static void draw_knob(lv_event_t * e);
 static bool is_slider_horizontal(lv_obj_t * obj);
 static void drag_start(lv_obj_t * obj);
 static void update_knob_pos(lv_obj_t * obj, bool check_drag);
+
+#if LV_USE_OBSERVER
+    static void slider_value_changed_event_cb(lv_event_t * e);
+    static void slider_value_observer_cb(lv_observer_t * observer, lv_subject_t * subject);
+#endif /*LV_USE_OBSERVER*/
 
 /**********************
  *  STATIC VARIABLES
@@ -215,6 +221,25 @@ bool lv_slider_is_symmetrical(lv_obj_t * obj)
 {
     return lv_bar_is_symmetrical(obj);
 }
+
+#if LV_USE_OBSERVER
+lv_observer_t * lv_slider_bind_value(lv_obj_t * obj, lv_subject_t * subject)
+{
+    LV_ASSERT_NULL(subject);
+    LV_ASSERT_NULL(obj);
+
+    if(subject->type != LV_SUBJECT_TYPE_INT && subject->type != LV_SUBJECT_TYPE_FLOAT) {
+        LV_LOG_WARN("Incompatible subject type: %d", subject->type);
+        return NULL;
+    }
+
+    lv_obj_add_event_cb(obj, slider_value_changed_event_cb, LV_EVENT_VALUE_CHANGED, subject);
+
+    lv_observer_t * observer = lv_subject_add_observer_obj(subject, slider_value_observer_cb, obj, NULL);
+    return observer;
+}
+#endif /*LV_USE_OBSERVER*/
+
 
 /**********************
  *   STATIC FUNCTIONS
@@ -636,5 +661,38 @@ static void update_knob_pos(lv_obj_t * obj, bool check_drag)
             return;
     }
 }
+
+
+#if LV_USE_OBSERVER
+
+static void slider_value_changed_event_cb(lv_event_t * e)
+{
+    lv_obj_t * slider = lv_event_get_current_target(e);
+    lv_subject_t * subject = lv_event_get_user_data(e);
+
+    if(subject->type == LV_SUBJECT_TYPE_INT) {
+        lv_subject_set_int(subject, lv_slider_get_value(slider));
+    }
+#if LV_USE_FLOAT
+    else {
+        lv_subject_set_float(subject, (float)lv_slider_get_value(slider));
+    }
+#endif
+}
+
+static void slider_value_observer_cb(lv_observer_t * observer, lv_subject_t * subject)
+{
+    if(subject->type == LV_SUBJECT_TYPE_INT) {
+        lv_slider_set_value(observer->target, subject->value.num, LV_ANIM_OFF);
+    }
+#if LV_USE_FLOAT
+    else {
+        lv_slider_set_value(observer->target, (int32_t)subject->value.float_v, LV_ANIM_OFF);
+    }
+#endif
+}
+
+#endif /*LV_USE_OBSERVER*/
+
 
 #endif
